+++ Demo/Shared/GitHubAPI.swift
@@ -52,6 +52,9 @@ extension GitHub: TargetType {
             return nil
         }
     }
+    public var parameterEncoding: ParameterEncoding {
+        return URLEncoding.default
+    }
     public var task: Task {
         return .request
     }